Reorganized files (but did not update import paths)
