+++ src/pages/auth/Signout.js
