+++ src/pages/discover/Incidents.js
