+++ src/pages/discover/Legend.js
