+++ src/pages/discover/PlaceDetails.js
