+++ src/pages/discover/PlacesResults.js
