+++ src/pages/discover/PlacesResultsFilter.js
