+++ src/pages/discover/SearchPlaceholder.js
